Update now loops through all files and updates them all per the specified update count
Before it would grab a random number of files and update them. Now the Update Count has been changed to Number of Updates, and it will take all files in the specified directory, update them all, then commit them all and it will do this for the number of times specified in the "Number of Updates" field.

The idea is, I can create a directory with 1 or a few files, then run a 1000 Number of Updates on those files to generate a very large number of revisions.